pyproject.toml
# Copyritht © 2023 Marcin Różewski MDSANIMA

[build-system]
requires = ['setuptools>=67.1.0', 'wheel>=0.38.4', 'pip>=23.0', 'setuptools_scm[toml]>=7.1']
build-backend = 'setuptools.build_meta'

[project]
name = 'mdsanima_cli'
authors = [{name = 'Marcin Różewski', email = 'marcinrozewski@gmail.com'}]
description = 'Command line tools for images processing, generating pixelart, adding logo to images and many more.'
readme = 'README.md'
license = {file = 'LICENSE'}
requires-python = '>=3.10'
keywords = [
    'cli',
    'command line',
    'console-app',
    'console-application',
    'exif metadata',
    'image-processing',
    'images',
    'mdsanima command line',
    'mdsanima-cli',
    'mdsanima-dev',
    'mdsanima',
    'meta information',
    'pixelart',
    'python-library',
    'python',
    'python3',
    'shell-script',
]
classifiers = [
    'Development Status :: 1 - Planning',
    'Environment :: Console',
    'Environment :: Plugins',
    'Intended Audience :: Developers',
    'Intended Audience :: Education',
    'Intended Audience :: End Users/Desktop',
    'Intended Audience :: Information Technology',
    'Intended Audience :: Legal Industry',
    'Intended Audience :: Other Audience',
    'Intended Audience :: System Administrators',
    'License :: OSI Approved :: Apache Software License',
    'Natural Language :: English',
    'Natural Language :: Polish',
    'Operating System :: MacOS',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: POSIX :: BSD',
    'Operating System :: POSIX :: Linux',
    'Operating System :: Unix',
    'Programming Language :: Python :: 3.10',
    'Topic :: Desktop Environment',
    'Topic :: Documentation :: Sphinx',
    'Topic :: Education :: Testing',
    'Topic :: Multimedia',
    'Topic :: Scientific/Engineering :: Image Processing',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: System :: Console Fonts',
    'Topic :: System :: Shells',
    'Topic :: System :: System Shells',
    'Topic :: Terminals',
    'Topic :: Text Processing',
    'Topic :: Utilities',
]
dependencies = [
    'Pillow==9.4.0',
    'mdsanima-dev==0.1.1',
]
dynamic = ['version']

[project.urls]
homepage = 'https://github.com/mdsanima-lab/mdsanima-cli'
documentation = 'https://github.com/mdsanima-lab/mdsanima-cli#readme'
repository = 'https://github.com/mdsanima-lab/mdsanima-cli.git'
changelog = 'https://github.com/mdsanima-lab/mdsanima-cli/blob/main/CHANGELOG.md'

[project.scripts]
mdsanima = 'mdsanima_cli:main_cli'

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = 'black'
src_paths = ['src', 'tests']
force_single_line = true
lines_before_imports = 2
lines_after_imports = 2
treat_all_comments_as_code = true
extend_skip = ['.md', '.json']

[tool.setuptools.packages.find]
where = ['src']

[tool.setuptools_scm]
write_to = 'src/mdsanima_cli/_version.py'
